Fix Skiko backed Haze not display multiple Rects
Caused the runtime shaders being drawn on top of each,
meaning that only the last rectangle is actually visible.

Fixed by updating the runtime effect to only draw the blurred
content, and transparent for anything else. ImageFilter.makeMerge
then draws each filter on top of each other.

Also updated the samples to include a bottom bar, validating that
this works.