Fix Skiko backed Haze not displaying multiple Rects (#17)
Caused the runtime shaders drawing the entire content on top
of each, meaning that only the last rectangle is actually visible.

Fixed by updating the runtime effect to only draw the blurred
content, and transparent for anything else. ImageFilter.makeMerge
then draws each filter on top of each other.

Also updated the samples to include a bottom bar, validating that
this works.