Apply noise texture using HARD_LIGHT
We currently use OVERLAY, which has the unfortunate effect
of changing the luminosity over the content.

After some experimentation and research, it seems that HARD_LIGHT
is a more aggresive relation to OVERLAY, which keeps black pixels black,
and similar for white. The results are much better to my eye.