[Android] Fix dialog being offset for non edge-to-edge windows
Currently our window offset logic assumes that the host content is drawn
at [0,0], which usually happens when content is drawn edge-to-edge.
There are still apps which don't use that, or tools which don't
(i.e. deploy previews), so we need to handle those cases too. The issue
why we don't is that we currently use the host ComposeView's screen location,
but that isn't quite correct in these situations, we need to use the window's
decor view.