[next] Initial upgrade to Compose Multiplatform 1.7.0-alpha (#250)
This is the initial PR to integrate Compose Multiplatform 1.7.0
(currently alpha) in Haze.

The big goal here is to merge the different platform implementations
into 1 common implementation powered by the new `GraphicsLayer` APIs in
Compose 1.7.0.

Since Compose GraphicsLayers are similar in concept and API to Android
RenderNodes, refactoring the existing Android implementation to work on
all platforms has been fairly easy. The `RenderEffect` applied is
different per platform, but all of the layer handling is in
`commonMain`.