New rendering mode (#259)
Currently Haze is all smoke and mirrors, as it draws all of the blurred
areas in the haze layout node. The hazeChild nodes just update the size,
shape, etc, which the haze modifier reads, to know where to draw.

This PR inverts this, so that the haze node creates a graphics layer and
puts it in the HazeState. Each hazeChild then uses that graphics layer
to apply the blurred effect locally in the relevant layout node.

Fixes #213, #229, #245, #256