Add workaround for invalidations not happening on Skia backed platforms (#296)
We need to re-draw children (`Modifier.hazeChild`) when the background
content `Modifier.haze`) has updated the `contentLayer`.

That works automatically on Android, as it uses `RenderNodes` which
automatically repaint when an upstream `RenderNode` is updated. The Skia
backed platforms do not do this though, so we can't rely on this
behavior.

This PR adds in a state backed `invalidationTick`, allowing the
`hazeChild` on Skia-backed platforms to know when the content layer has
been updated. Why don't we just the `contentLayer` in a state I hear you
ask? GraphicsLayers aren't state. They don't implmenent equality, so
every draw will trigger readers that the value has changed. This means
that we can easily get into invalidation loops:

1. Parent draws, creates a new GraphicsLayer and updates state.
2. Children are notified that the GL has changed. They trigger a draw
invalidation.
3. Parent is drawn, creates a new GraphicsLayer and updates state.
4. Repeat.

The invalidationTick added here would act exactly the same, but because
it's an `Int` we can use it to know when NOT to invalidate in step 2.
This is the very giant hack in this MR: we only invalidate every other
invalidation tick change.