Add support for multiple Haze nodes attached to a HazeState (#441)
This PR adds in support for multiple `Modifier.haze` nodes attached to a
`HazeState`. This means that there can be N haze nodes, and M hazeChild
nodes attached to a state.

This has a number of use cases which I'll document, but the main one is
the overlapping area use case:


https://github.com/user-attachments/assets/e7c71898-4988-47f1-b094-c43e6bce7422

- [x] Benchmark performance
- [x] Write docs

Fixes #425, #443