[Skia] Fix hazeSource content changes not updating hazeEffects (#476)
There's two related issues which this PR fixes:

- We no longer draw graphics layers in a
`Snapshot.withoutReadObservation`. This was done to avoid unwanted
invalidations in Haze, but it has the knock-on effect that it breaks
internal invalidation in the Skiko GraphicsLayer impl. Fixed by using
`withoutReadObservation` more locally around state reads.
- We were not observing `HazeArea.contentLayer`. This probably isn't the
root cause of this issue, but this is more defensive for the future.

Fixes #475