Re-introduce expanded layer size (#522)
This PR introduces the expanded layer size, which was removed in #461. 

The key difference in this implementation is that we now no longer
expand the layer size beyond what the content bounds. This means that
the edges are now either blurred using the expanded content, or the
clamp edge treatment. The previous implementation would not allow the
clamp edge treatment to work.

Fixes #516