Fix flicker when using AndroidX Navigation (#566)
Caused by use observing the `LocalLifecycleOwner` to determine when to
manually clear the graphics layer when the Activity is stopped.

However, when using AndroidX Navigation the `LocalLifecycleOwner` is
tied to the back stack entry, which isn't what we really want to
observe. Fixed by just observing the Activity instead.

Fixes #559