Add tests which illustrate problem in horizontal pager
When a page in a pager is more than halfway off the screen, the blurring
assumes the content is not intersecting with the screen and does not
render it.

This adds several tests of the credit card inside a pager to demonstrate
where the problem occurs (when pageOffsetFraction is negative) and when
it doesn't.