Add RenderScript backed blur implementation (#590)
This PR adds in a RenderScript backed blur implementation.

As part of the changes, we now have a mode which allows us to
automatically invalidate `hazeEffect` nodes from pre-draw. While this is
necessary for the RenderScript implementation, it **also** fixes the
long standing issue which resulting in me disabling support on Android
12 (API Level 31).

Docs and performance results will come in a follow-up PR.

Fixes #77 
Fixes #140

---------

Co-authored-by: Serhii Yaremych <sergeyfitis@gmail.com>