Noise fixes for Android
For RenderScript, we no longer scale the noise or
tints at all. They get drawn over the blurred content
at the original size.

For RenderEffect, the noise is applied as part of the
RenderEffect which makes it tricky in terms of scaling.
I've disable scaling for now, meaning that input scaled
content will have a 'larger' (in 2d space) noise effect.
It shouldn't matter for most usages.