Stopping clipping the drawn layer size
We've used the 'hack' of using an expanded sized layer, to
help combat visual effects of the CLAMP tile mode
for a while.

Recently I added an optimization to clip this layer to the
size of the 'source' areas, but this has had a negative effect
on certain situations, primarily when high contrasting content
is entering (or exiting) the screen.

This PR removes the optimization, and moves us back to what
we had around Haze 1.0.

Fixes #653