Fix overlapping content with layer transformations (#674)
There was a small optimization I made in the blurring code, to reject
areas which do not overlap, according to their layout bounds. Layer
transformations aren't based on layout sizes though, so we can not
safely assume that the overlapping check works correctly.

It's a tiny optimization, so better to just remove it.

Fixes #675